Make priv generation into a compiler task
