+++ lib/livebook_web/endpoint.ex
@@ -33,17 +33,17 @@ defmodule LivebookWeb.Endpoint do
   # To account for both cases, we configure Plug.Static :from as MFA
   # and return the accessible priv/ location in both scenarios.
   #
-  # The priv/ static files are generated by the livebook.gen_priv task
-  # before building the escript or the release. We gzip the static
-  # files in priv/, since we want to serve them gzipped, and we don't
-  # include the non-gzipped ones to minimize app size. Note that we
-  # still have a separate static/ directory with the CI-precompiled
-  # assets, which we keep in Git so that people can install escript
-  # from GitHub or run MIX_ENV=prod phx.server, without Node and NPM.
-  # Storing minified assets is already not ideal, but we definitely
-  # want to avoid storing the gzipped variants in Git. That's why we
-  # store the assets uncompressed and then generate priv/static with
-  # their compressed variants as part of the build process.
+  # The priv/ static files are generated by the compile.livebook_priv
+  # as part of compilation. We gzip the static files in priv/, since
+  # we want to serve them gzipped, and we don't include the non-gzipped
+  # ones to minimize app size. Note that we still have a separate
+  # static/ directory with the CI-precompiled assets, which we keep
+  # in Git so that people can install escript from GitHub or run
+  # MIX_ENV=prod phx.server, without Node and NPM. Storing minified
+  # assets is already not ideal, but we definitely want to avoid
+  # storing the gzipped variants in Git. That's why we store the
+  # assets uncompressed and then generate priv/static with their
+  # compressed variants at compile time.
 
   if code_reloading? do
     # In development, we use assets from tmp/static_dev, which are
